Refactor bucket lifecycle methods to seperate file - api-get-lifecycle.go
