+++ api-get-policy.go
@@ -76,53 +76,3 @@ func (c Client) getBucketPolicy(bucketName string) (string, error) {
 	policy := string(bucketPolicyBuf)
 	return policy, err
 }
-
-// GetBucketLifecycle - get bucket lifecycle.
-func (c Client) GetBucketLifecycle(bucketName string) (string, error) {
-	// Input validation.
-	if err := s3utils.CheckValidBucketName(bucketName); err != nil {
-		return "", err
-	}
-	bucketLifecycle, err := c.getBucketLifecycle(bucketName)
-	if err != nil {
-		errResponse := ToErrorResponse(err)
-		if errResponse.Code == "NoSuchLifecycleConfiguration" {
-			return "", nil
-		}
-		return "", err
-	}
-	return bucketLifecycle, nil
-}
-
-// Request server for current bucket lifecycle.
-func (c Client) getBucketLifecycle(bucketName string) (string, error) {
-	// Get resources properly escaped and lined up before
-	// using them in http request.
-	urlValues := make(url.Values)
-	urlValues.Set("lifecycle", "")
-
-	// Execute GET on bucket to get lifecycle.
-	resp, err := c.executeMethod(context.Background(), "GET", requestMetadata{
-		bucketName:  bucketName,
-		queryValues: urlValues,
-	})
-
-	defer closeResponse(resp)
-	if err != nil {
-		return "", err
-	}
-
-	if resp != nil {
-		if resp.StatusCode != http.StatusOK {
-			return "", httpRespToErrorResponse(resp, bucketName, "")
-		}
-	}
-
-	bucketLifecycleBuf, err := ioutil.ReadAll(resp.Body)
-	if err != nil {
-		return "", err
-	}
-
-	lifecycle := string(bucketLifecycleBuf)
-	return lifecycle, err
-}
